Add functionality for storing cmd options Doom supported
Create structs to store options that could be passed into
Doom. Create method to parse the options from the command
line and store it into the struct. The way it currently works
is it looks for a cmd arg starting with a -, iterates past it
to grab the values for the option until it encounters another
arg that starts with -. Rinse and repeat until we reach
the end of the the args.

I might have missed some options but they will be added as I
find them as I implement more of the codebase.

IMPLEMENTS
m_argv.h/m_argv.c

There are some extra features such as not
segfaulting when an invalid/not enough option values are passed in.

TODO
Currently we don't support the @responsefile arg yet, as that
requires some more thinking on my end how to implement.

We should also be on our way to replace
all the bool flags found throughout the original code base as well.

Think of a way to define the value type for each option instead
of storing the values in a string seperated by spaces