Remove uneeded check in set_options
We were doing a check to see if the first option passed into
the set_options method was valid. An option is valid if it starts with a -.
If it doesn't the program should panic.

This is an uneeded check because if the first option
is not valid, the call to get_option_by_name_mut
will return None and we will still panic because of an invalid
option being passed in.

Updated the test name to be more generic and test for an invalid
option.