Remove uneeded cmd_args.is_empty() check in set_options
At the beginning of set_options we were doing a check
to see if the cmd_args passed into the program had
any elements in it. If it didn't, we returned from the function.

This is uneeded as the main way we loop through cmd_args and
process it is by comparing the index we are currently processing to
the length of cmd_args. If cmd_args length is zero, the while
loop will never be triggered.

Added assertions to the no_cmd_args test to make
sure none of the options are enabled or have a value set