Refactor DoomOptions to be created from cmd_args directly
Instead of setting the options based on the command line arguments
via a method in main, we are now instead creating a DoomOptions
instance via the new method with the cmd_args passed in that.
Then, set_options is called in the new method to do our
cmd_args checks. This makes it so a person just needs
to pass in the cmd_args to create the struct and it allows
us to set our DoomOptions struct as nonmutable instead
of mutable