Refactor creation of DoomOptions with default option values
Doing the struct intialization of DoomOption inside the DoomOptions
new method was taking up a lot of lines. I instead created a const
array of tuple data for every option with have so far. Takes
up a lot less lines and can be used to validate that the DoomOptions
struct is being initialized with all the options in the one unit test