Add DoomOption unit test for invalid response file
Added a unit test that tests for the app to panic
if it can't find a response file. I'm adding this
because while trying it out in powershell, I discovered
@ is treated as an operator, which caused the response
file path to not be passed in when it was just @word with
no slashes. Thought the functionality was broken for way longer than
I care to admit lol.