Refactor set options
Refactor set options to not need get_option_by_name_mut. Found an easier
way to get the mut reference I needed. That was the
only place it was used so it was removed along with the unit test.

Rename set_options to create_options and made it so creating the inital
doom option vec was done in that function instead of the new function