Remove DoomOptions.is_dev_option_enabled method
After looking at the original source more closely,
the -shdev, -regdev, and -comdev options are not used after
the initialization of config values at the beginning of the program.
Although in the original source code each of the options are
setting the dev param flag to true, the code later goes on to check
if the -devparm is explicilty set and sets the flag to whether it
is or not.

This means where the -shdev, -regdev, and -comdev
flags are setting the devparm bool is completly useless. Since
this is the case, we don't need a method to check if one of the dev
params are set and will check if -devparam is set using the
is_option_enabled method when it is needed.