Refactor DoomOption values to be Vec<String>
We are at a point in time in the original code base where the game
starts asking for options that have values associated with them.
Since the values of DoomOption was multiple values in a string
seperated by a space, it would get difficult to manipulate multiple
values. So I changed it be a Vec<String>. Since we are changing
how values are stored, I had to change the definition of enabled.

I removed the enabled method and made it a flag in DoomOption. The
reason for this is some options(Such as -shdev) don't have any values
associated with them. I originally added a "true" value for such
options, but that could get ugly later on down the line. So now wether
an option is enabled or not is no longer determined by the presence of
a value in DoomOption.value. Now we are setting enabled in the new
method based on if it was passed in and it meets the min/max values
needed for it.

Refactored unit tests for these changes. Added a helper method in the
test module to construct a cmd_args Vec from a Hashmap where the key
is the option name and the value is an array slice of values. This
makes testing that the correct values are set for an option much easier
and reduces the number of lines of code we have in this file.