Add logic to process wad and lump files
Add logic to process wad/lump files. This logic will read
a wad/lump file, and extract the relevant lump data that we
need to access the data later(Name, data size, location in file,
file it originated from).

This logic will panic if:
        An unsupported WAD ID is found in a wad(Not IWAD or PWAD)
        A lump file name is > 8 characters
All other issues will be wrapped into a WadError struct
and returned and displayed to the user. In this case, the
game will continue processing wads and ignore the wad with
the error. This is how the original linux source behaved.

Lumps found in wads/lumps are stored in the order they are processed.
If a lump with the same name gets processed twice from different files,
the most recent one should overrun the previous instances

This logic replicates W_AddFile/W_InitMultipleFiles(w_wad.c)
in the original source