Refactor unit tests for option.rs
Fixed unit tests for option.rs to help with the
test coverage.

Removed the error handling in the tests and replaced it with
unwrap. Usually, unwrap is really bad since it panics but
thats what we want to happen if a scenario isn't met in a unit
test. Using unwrap gets rid of unecessary code and helps with
tests coverage since llvm-cov marks things like unwrap_or_else
or panic in tests as uncovered

Added test case for responsefile to check that it will ignore lines
in it if it contains non ascii characters(such as Japanese)