Fix error with reload prefix on ~test.wad format
Fixed error when processing a file and determining if it
was a reloadable file or not. For some reason, path_buffer.starts_with
doesn't work in the format ~test.wad, or a wad named test.wad in
the current directory. Switched the starts_with check to be
on the string version of the path and it works now