Abstract lumps into DoomFile struct(wad.rs)
Abstracted the lumps into a DoomFile struct. This struct
will represent a file(.wad/.lmp) and its associated lumps.
It will also contain a file handle to an open file representing the
wad/lmp in the OS. This way I can support loading from a already open
file/ reload the file and override the file handle when it's detected
to be a reloadable file. I would put the handle on the lump,
but I ran into issues with too many handles being open at once....